requirements.txt
azure-storage-blob
pyyaml
numpy
pandas